fix(clippy): cargo clippy --fix auto-applied suggestions across workspace
Machine-applicable lint fixes: map_or, sort_by_key/sort_unstable_by_key,
redundant closures, is_multiple_of, str::repeat, clamp, Error::other,
manual ASCII comparisons, redundant field names, etc. No behaviour change.

+++ services/expresso-mail/src/lmtp.rs
@@ -347,7 +347,7 @@ mod tests {
 
     #[test]
     fn max_msg_bytes_exceeds_max_rcpts() {
-        assert!(super::MAX_MSG_BYTES > super::MAX_RCPTS as usize);
+        assert!(super::MAX_MSG_BYTES > super::MAX_RCPTS);
     }
 
     #[test]
